training/requirements.txt
# Use versions that provide pre-compiled wheels for Windows/Python 3.10
tensorflow-gpu==2.10.0
tensorflowjs==3.21.0
numpy>=1.21.2
scipy>=1.7.3
opencv-python>=4.5.0
Pillow>=9.0.0
supabase>=2.3.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.0